+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Increase spacing between element in `ChatMessage`
 * Increase font size of `View More` text in `KeyValues`
 * Increase font size for iOS in `Select`
+* Support for no overlay and height control from parent in `ContainerOpenable`
 
 ### Fixed